Support Azure identity token in docker config
When using `az acr login` to authenticate with the ACR (Azure Container
Registry) the Docker `config.json` file gets populated with an `auth`
field that contains the base64 encoding of `username:password` where
`username` is `00000000-0000-0000-0000-000000000000` (that actually gets
ignored I believe) and `password` is empty, and an `identitytoken` entry.
The latter is a short-lived token and for the use case here can be used
as the password.

Tested with my own private ACR and with this change I was able to
`oci_pull` an image from it after having `az acr login` populate my
`~/.docker/config.json`.